requirements.txt
aerich==0.8.2
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
asyncpg==0.30.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
dictdiffer==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
loguru==0.7.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypika-tortoise==0.5.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.13.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
